[REF] *: rename 'menuItem' files to 'action' files
Following the previous type rename, this commit also renames the
files and also changes their location in the project organization
