+++ src/registries/menus/col_menu_registry.ts
@@ -1,10 +1,10 @@
+import * as ACTION_DATA from "../../actions/data_actions";
+import * as ACTION_EDIT from "../../actions/edit_actions";
+import * as ACTION_FORMAT from "../../actions/format_actions";
+import * as ACTION_INSERT from "../../actions/insert_actions";
+import * as ACTION_VIEW from "../../actions/view_actions";
 import { _lt } from "../../translation";
 import { MenuItemRegistry } from "../menu_items_registry";
-import * as ACTION_DATA from "./items/data_menu_items";
-import * as ACTION_EDIT from "./items/edit_menu_items";
-import * as ACTION_FORMAT from "./items/format_menu_items";
-import * as ACTION_INSERT from "./items/insert_menu_items";
-import * as ACTION_VIEW from "./items/view_menu_items";
 
 export const colMenuRegistry = new MenuItemRegistry();
