+++ src/registries/menus/index.ts
@@ -1,7 +1,6 @@
 export * from "./cell_menu_registry";
 export * from "./col_menu_registry";
 export * from "./link_menu_registry";
-export * from "./menu_items_actions";
 export * from "./number_format_menu_registry";
 export * from "./row_menu_registry";
 export * from "./sheet_menu_registry";
